meta(): reorganizes mocks into module
