+++ mockserver/mockserver/urls.py
@@ -14,12 +14,11 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path, include, re_path
+from django.urls import path, include
 
-from api.views import fetch_mock
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/v1/', include('api.urls')),
-    re_path(r"^.*/$", fetch_mock)
+    path('', include('mocks.urls')),
 ]
